movie_collections/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.1
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environ==1.0
factory-boy==3.3.0
Faker==27.0.0
idna==3.7
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
